This patch is intended to fix #133. It adds the computation of the core at every memory access when the access is shown to be within bounds, however, currently it changes the subsumption profile, especially for examples `arraysimple6.c` and `regexp_nonrecursive2.c`. of [tracer-x/klee-examples](https://gitlab.com/tracer-x/klee-examples/basic). The reason is that this patch introduces more components into the interpolant, some of which uses absolute memory addressing. Local values with absolute addressing is problematic as their absolute addresses are different between calls, yet we want to compare them for subsumption checks. This results in failure of subsumptions between different calling instances of the same function. A fix should devised that change the absolute addressing to relative (allocation site + offset) addressing.