@domainexpert 2016-09-14
I have made some attempts to implement a mechanism that compare only offsets at subsumption check, but this cannot solve the problem as pointers are still stored in the memory as absolute values, so when the contents are compared in a subsumption check, if the contents were pointers, then the their absolute values are compared, likely resulting in subsumption failure across calls to the same function.