@domainexpert 2016-10-28
I ran some experiments today, and following is the result.
**BEFORE PATCH**
- In `klee-examples/basic`:
  - **make bubble.cov** time: 0.16 s, coverage: 14 of 23 lines.
  - **make bubble.stpcov** time: 1.27s, coverage: 14 of 23 lines.
- In `klee-examples/scalability`:
  - **make statemate.cov** time: 0.16s, coverage: 349 of 1230 lines.
  - **make statemate.stpcov** time: 56.01s, coverage: 349 of 1230 lines.
  - **make Regexp.cov** time: 0.97s, coverage: 30 of 36 lines.
  - **make Regexp.stpcov** time: 6.65s, coverage: 31 of 36 lines.
**AFTER PATCH**
- In `klee-examples/basic`:
  - **make bubble.cov** time: 0.16 s, coverage: 14 of 23 lines.
  - **make bubble.stpcov** time: 1.24s, coverage: 14 of 23 lines.
- In `klee-examples/scalability`:
  - **make statemate.cov** time: 0.16s, coverage: 349 of 1230 lines.
  - **make statemate.stpcov** time: 55.56s, coverage: 349 of 1230 lines.
  - **make Regexp.cov** time: 1.60s, coverage: 30 of 36 lines.
  - **make Regexp.stpcov** time: 6.40s, coverage: 31 of 36 lines.
So here we are paying some performance penalty in the case of `Regexp`.