@domainexpert 2016-10-28
This is merged now as this is a reasonable patch when memory safety checks are done, although having some performance penalty.